\name{caribou-package}
\alias{caribou-package}
\alias{caribou}
\docType{package}
\title{  Estimation of caribou abundance }
\description{
   This package implements the methodology found in the article by Rivest, L.-P., Couturier, S. and Crepeau, H. (1998) about caribou abundance estimation. It also includes a function based on the Lincoln-Petersen Index as applied to radio-telemetry data by White and Garrott (1990).
}
\details{
\tabular{ll}{
Package: \tab caribou\cr
Type: \tab Package\cr
Version: \tab 0.3-0\cr
Date: \tab 2011-11-17\cr
License: \tab GPL-2\cr
} }
\author{ Helene Crepeau \email{Helene.Crepeau@mat.ulaval.ca} and\cr
Louis-Paul Rivest \email{Louis-Paul.Rivest@mat.ulaval.ca} and\cr
Sophie Baillargeon \email{Sophie.Baillargeon@mat.ulaval.ca} }
\references{
Rivest, L.-P., Couturier, S. and Crepeau, H. (1998). Statistical Methods for estimating caribou abundance using postcalving aggregations detected by radio telemetry. \emph{Biometrics}, \bold{54}(3), 865-876.

White, G.C. and Garrott, R.A. (1990). \emph{Analysis of wildlife radio-tracking data}. San Diego: Academic Press.
}
\keyword{ package }
